archive/services/nxt-brain/requirements.txt
fastapi
rich
python-dotenv
uvicorn
pydantic
httpx
